Creation of the object-oriented structure ‘project’
Object-oriented structure that permits to store and then to write the
model outputs of a same project in files. A project can include several
scenarios. It will contain all the methods to write outputs in excel or
word documents. It is pretty empty for the moment but we will populate
this new class little by little, until perhaps having an automatic
report writer.