Small debug to model runner
Not sure why this code ("and scenario == 0") was included, but the associated bug prevents running of scenarios under the epidemiological uncertainty approach. This possibly isn't a good idea anyway, but at least by removing it, we can do this if we want to. Still bugs in outputs if you request results from the scenarios.